Gave the ability to have zonal summary use cached results.
- Added a .cached method to DataSourceLike that caches the results
  of the elements on the local machine.
- Refactored zonal summary methods to take in cached results
  to be used on the full tile intersections.
