+++ src/main/scala/geotrellis/raster/op/zonal/summary/TileSummary.scala
@@ -1,0 +1,10 @@
+package geotrellis.raster.op.zonal.summary
+
+import geotrellis._
+import geotrellis.source._
+
+trait TileSummary[T,V,That <: DataSource[_,V]] {
+  def handlePartialTile[D](pt:PartialTileIntersection[D]):T
+  def handleFullTile(pt:FullTileIntersection):T
+  def converge(ds: DataSource[T,_]):That
+}
