+++ src/main/scala/geotrellis/source/DataSource.scala
@@ -12,3 +12,21 @@ import geotrellis.statistics._
   */
 trait DataSource[+T,+V] extends DataSourceLike[T,V,DataSource[T,V]] {
 }
+
+object DataSource {
+  def convergeSeq[A](elementOps:(Op[Seq[Op[A]]])) = {
+    logic.Collect(elementOps)
+  }
+
+  def fromValues[T](elements:T*):DataSource[T,Seq[T]] =
+    fromValues(elements)
+
+  def fromValues[T](elements:Seq[T])(implicit d:DI):DataSource[T,Seq[T]] =
+    apply(Literal(elements.map(Literal(_))))
+
+  def apply[T](elements:Op[Seq[Op[T]]]):DataSource[T,Seq[T]] = {
+    val builder:DataSourceBuilder[T,Seq[T]] = new DataSourceBuilder(convergeSeq)
+    builder.setOp(elements)
+    builder.result
+  }
+}
